회원가입 시 사용자를 판매자 그룹에 추가 및 판매자&일반 페이지 구분
